+++ src/pages/planos/index.tsx
@@ -2,12 +2,23 @@ import type { PageProps } from 'gatsby'
 import { graphql } from 'gatsby'
 import type { PlanosQuery } from '@generated/graphql'
 import { mark } from 'src/sdk/tests/mark'
+import SimpleText from 'src/components/sections/SimpleText/SimpleText'
 import Section from 'src/components/sections/Section'
 
 export type Props = PageProps<PlanosQuery>
+function Page(props: Props) {
+  const {
+    data: { allContentfulPlanosTextoSimples },
+  } = props
 
-function Page() {
-  return <Section />
+  return (
+    <Section>
+      <SimpleText
+        textReceived={allContentfulPlanosTextoSimples}
+        className="text-banner-bottom"
+      />
+    </Section>
+  )
 }
 
 export const querySSG = graphql`
@@ -17,6 +28,13 @@ export const querySSG = graphql`
         title
         description
         titleTemplate
+      }
+    }
+    allContentfulPlanosTextoSimples {
+      nodes {
+        text {
+          text
+        }
       }
     }
   }